+++ packages/core/story-structure.js
@@ -54,8 +54,11 @@ const s = [
               'c/Datagrid/Extensions/EditableCell',
               'c/Datagrid/Extensions/ColumnCustomization',
               'c/Datagrid/Extensions/Skeleton',
+              'c/Datagrid/Extensions/Slug',
             ],
           },
+          { n: 'TruncatedList', s: ['c/TruncatedList'] },
+          { n: 'DelimitedList', s: ['c/DelimitedList'] },
         ],
       },
       {
@@ -106,6 +109,10 @@ const s = [
           { n: 'Notifications', s: ['c/NotificationsPanel'] },
           { n: 'Remove', s: ['c/RemoveModal'] },
           { n: 'Saving', s: ['c/Saving'] },
+          {
+            n: 'Search bar',
+            s: ['c/SearchBar', 'c/SearchBarScopes'],
+          },
           { n: 'Status icons', s: ['c/StatusIcon'] },
           { n: 'Toolbars', s: ['c/Toolbar'] },
           { n: 'User profile images', s: ['c/UserProfileImage'] },
@@ -135,8 +142,33 @@ const s = [
       {
         n: 'Novice to pro',
         s: [
+          {
+            n: 'Coachmark',
+            s: [
+              'c/Coachmark',
+              'c/CoachmarkFixed',
+              'c/CoachmarkStack',
+              {
+                n: 'Coachmark Elements',
+                s: [
+                  'c/CoachmarkBeacon',
+                  'c/CoachmarkButton',
+                  'c/CoachmarkOverlayElement',
+                  'c/CoachmarkOverlayElements',
+                ],
+              },
+            ],
+          },
           'c/Checklist',
           'c/Guidebanner',
+          {
+            n: 'Interstitial Screen',
+            s: [
+              'c/InterstitialScreen',
+              'c/InterstitialScreenView',
+              'c/InterstitialScreenViewModule',
+            ],
+          },
           'c/InlineTip',
           'c/NonLinearReading',
         ],
@@ -148,11 +180,9 @@ const s = [
     s: [
       {
         n: 'Components',
-        s: [
-          { n: 'Example', s: ['c/ExampleCommunityComponent'] }
-        ]
-      }
-    ]
+        s: [{ n: 'Example', s: ['c/ExampleCommunityComponent'] }],
+      },
+    ],
   },
   {
     n: 'Carbon',
@@ -172,14 +202,14 @@ const getSectionOrder = (sArray) =>
     .map((entry) =>
       typeof entry === 'string'
         ? // if the entry is a string, return it (without the prefix)
-        getEntryDisplayName(entry)
+          getEntryDisplayName(entry)
         : // if the entry is another structure, return its name, but first get
-        // the fully materialized paths it contains and add the entry name
-        // to the front of each
-        [
-          getSectionOrder(entry.s).map((path) => `${entry.n}/${path}`),
-          entry.n,
-        ]
+          // the fully materialized paths it contains and add the entry name
+          // to the front of each
+          [
+            getSectionOrder(entry.s).map((path) => `${entry.n}/${path}`),
+            entry.n,
+          ]
     )
     .flat(Infinity);
 
@@ -212,13 +242,13 @@ const getPath = (s, componentAndPrefix, componentName) =>
       found ||
       (typeof next === 'string'
         ? // if this entry is a string, and it matches the component name
-        // we're looking for, return it as an array, else return null
-        getEntryPrefixAndComponentName(next) === componentAndPrefix
+          // we're looking for, return it as an array, else return null
+          getEntryPrefixAndComponentName(next) === componentAndPrefix
           ? [getEntryDisplayName(next)]
           : null
         : // if this entry is another structure, find the materialized path
-        // into it and prepend its name if found and return null otherwise
-        prepend(next.n, getPath(next.s, componentAndPrefix, componentName))),
+          // into it and prepend its name if found and return null otherwise
+          prepend(next.n, getPath(next.s, componentAndPrefix, componentName))),
     null
   );