+++ packages/ibm-products/src/components/Datagrid/Datagrid/addons/Filtering/FilterPanel.js
@@ -5,8 +5,6 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-/* eslint-disable react/jsx-key */
-
 import { Accordion, AccordionItem, Button, Layer, Search } from '@carbon/react';
 import { BATCH, CLEAR_FILTERS, INSTANT, PANEL } from './constants';
 import React, { useContext, useEffect, useMemo, useRef, useState } from 'react';
@@ -23,7 +21,7 @@ import {
 
 import { ActionSet } from '../../../../ActionSet';
 import { Close } from '@carbon/react/icons';
-import { FilterContext } from '.';
+import { FilterContext } from './FilterProvider';
 import PropTypes from 'prop-types';
 import cx from 'classnames';
 import { motion, useReducedMotion } from 'framer-motion';
@@ -35,10 +33,19 @@ export const componentClass = `${blockClass}-filter-panel`;
 
 const MotionActionSet = motion(ActionSet);
 
+const defaults = {
+  title: 'Filter',
+  closeIconDescription: 'Close filter panel',
+  primaryActionLabel: 'Apply',
+  secondaryActionLabel: 'Cancel',
+  searchLabelText: 'Filter search',
+  searchPlaceholder: 'Find filters',
+};
+
 const FilterPanel = ({
-  title = 'Filter',
-  closeIconDescription = 'Close filter panel',
-  updateMethod = BATCH,
+  title = defaults.title,
+  closeIconDescription = defaults.closeIconDescription,
+  updateMethod,
   filterSections,
   setAllFilters,
   onApply = () => {},
@@ -47,10 +54,10 @@ const FilterPanel = ({
   onPanelClose = () => {},
   showFilterSearch = false,
   filterPanelMinHeight = 600,
-  primaryActionLabel = 'Apply',
-  secondaryActionLabel = 'Cancel',
-  searchLabelText = 'Filter search',
-  searchPlaceholder = 'Find filters',
+  primaryActionLabel = defaults.primaryActionLabel,
+  secondaryActionLabel = defaults.secondaryActionLabel,
+  searchLabelText = defaults.searchLabelText,
+  searchPlaceholder = defaults.searchPlaceholder,
   reactTableFiltersState = [],
   autoHideFilters = false,
   isFetching = false,
@@ -191,9 +198,12 @@ const FilterPanel = ({
     const actionSetHeight =
       actionSetRef.current?.getBoundingClientRect().height;
 
-    const height = `calc(100vh - ${filterHeadingHeight}px - ${
-      showFilterSearch ? filterSearchHeight : 0
-    }px - ${updateMethod === BATCH ? actionSetHeight : 0}px)`;
+    const height = panelOpen
+      ? `calc(100vh - ${filterHeadingHeight}px - ${
+          /* istanbul ignore next */
+          showFilterSearch ? filterSearchHeight : 0
+        }px - ${updateMethod === BATCH ? actionSetHeight : 0}px)`
+      : 0;
 
     return height;
   };
@@ -229,6 +239,7 @@ const FilterPanel = ({
             onClick={closePanel}
           />
           {showFilterSearch && (
+            /* istanbul ignore next */
             <div ref={filterSearchRef} className={`${componentClass}__search`}>
               <Layer>
                 <Search