+++ packages/ibm-products/src/components/Datagrid/Extensions/Filtering/Panel.stories.js
@@ -1,10 +1,11 @@
-/* eslint-disable react-hooks/exhaustive-deps */
 /**
- * Copyright IBM Corp. 2022, 2023
+ * Copyright IBM Corp. 2022, 2024
  *
  * This source code is licensed under the Apache-2.0 license found in the
  * LICENSE file in the root directory of this source tree.
  */
+
+/* eslint-disable react-hooks/exhaustive-deps */
 
 import React, { useState } from 'react';
 import { Tooltip } from '@carbon/react';
@@ -27,7 +28,6 @@ import { DocsPage } from './Filtering.docs-page';
 import { StatusIcon } from '../../../StatusIcon';
 import { action } from '@storybook/addon-actions';
 import { makeData } from '../../utils/makeData';
-import { pkg } from '../../../../settings';
 import styles from '../../_storybook-styles.scss';
 
 export default {
@@ -46,7 +46,7 @@ export default {
       },
     },
   },
-  excludeStories: ['FilteringUsage', 'filterProps'],
+  excludeStories: ['FilteringUsage', 'filterProps', 'getDateFormat'],
 };
 
 // This is to show off the View all button in checkboxes
@@ -160,17 +160,30 @@ export const FilteringUsage = ({ defaultGridProps }) => {
     useColumnCenterAlign
   );
 
-  // Warnings are ordinarily silenced in storybook, add this to test
-  pkg._silenceWarnings(false);
-  // Enable feature flag for `useFiltering` hook
-  pkg.feature['Datagrid.useFiltering'] = true;
-  pkg._silenceWarnings(true);
-
   return <Datagrid datagridState={datagridState} />;
 };
 
 const FilteringTemplateWrapper = ({ ...args }) => {
   return <FilteringUsage defaultGridProps={{ ...args }} />;
+};
+
+// Example usage of mapping locale to flatpickr date format or placeholder value (m/d/Y or mm/dd/yyyy)
+export const getDateFormat = (lang, full) => {
+  const formatObj = new Intl.DateTimeFormat(lang).formatToParts(new Date());
+  return formatObj
+    .map(({ type, value }) => {
+      switch (type) {
+        case 'day':
+          return full ? 'dd' : 'd';
+        case 'month':
+          return full ? 'mm' : 'm';
+        case 'year':
+          return full ? 'yyyy' : 'Y';
+        default:
+          return value;
+      }
+    })
+    .join('');
 };
 
 export const filterProps = {
@@ -193,16 +206,18 @@ export const filterProps = {
             props: {
               DatePicker: {
                 datePickerType: 'range',
+                locale: navigator?.language || 'en',
+                dateFormat: getDateFormat(navigator?.language || 'en'),
               },
               DatePickerInput: {
                 start: {
                   id: 'date-picker-input-id-start',
-                  placeholder: 'mm/dd/yyyy',
+                  placeholder: getDateFormat(navigator?.language || 'en', true),
                   labelText: 'Joined start date',
                 },
                 end: {
                   id: 'date-picker-input-id-end',
-                  placeholder: 'mm/dd/yyyy',
+                  placeholder: getDateFormat(navigator?.language || 'en', true),
                   labelText: 'Joined end date',
                 },
               },
@@ -351,7 +366,6 @@ export const PanelInstant = prepareStory(FilteringTemplateWrapper, {
     filterProps: ARG_TYPES.filterProps,
   },
   args: {
-    featureFlags: ['Datagrid.useFiltering'],
     gridTitle: 'Data table title',
     gridDescription: 'Additional information if needed',
     useDenseHeader: false,
@@ -514,7 +528,6 @@ export const PanelWithInitialFilters = prepareStory(FilteringTemplateWrapper, {
     filterProps: ARG_TYPES.filterProps,
   },
   args: {
-    featureFlags: ['Datagrid.useFiltering'],
     initialState: {
       filters: [
         {
@@ -710,7 +723,6 @@ export const PanelOnlyAccordions = prepareStory(FilteringTemplateWrapper, {
     filterProps: ARG_TYPES.filterProps,
   },
   args: {
-    featureFlags: ['Datagrid.useFiltering'],
     gridTitle: 'Data table title',
     gridDescription: 'Additional information if needed',
     useDenseHeader: false,
@@ -874,7 +886,6 @@ export const PanelNoAccordions = prepareStory(FilteringTemplateWrapper, {
     filterProps: ARG_TYPES.filterProps,
   },
   args: {
-    featureFlags: ['Datagrid.useFiltering'],
     gridTitle: 'Data table title',
     gridDescription: 'Additional information if needed',
     useDenseHeader: false,
@@ -1038,7 +1049,6 @@ export const PanelNoData = prepareStory(FilteringTemplateWrapper, {
     filterProps: ARG_TYPES.filterProps,
   },
   args: {
-    featureFlags: ['Datagrid.useFiltering'],
     data: [],
     gridTitle: 'Data table title',
     gridDescription: 'Additional information if needed',
@@ -1202,7 +1212,6 @@ export const PanelManyCheckboxes = prepareStory(FilteringTemplateWrapper, {
     filterProps: ARG_TYPES.filterProps,
   },
   args: {
-    featureFlags: ['Datagrid.useFiltering'],
     gridTitle: 'Data table title',
     gridDescription: 'Additional information if needed',
     useDenseHeader: false,