+++ packages/ibm-products/src/components/index.js
@@ -20,6 +20,7 @@ export {
 export { CreateTearsheetNarrow } from './CreateTearsheetNarrow';
 export {
   EmptyState,
+  EmptyStateV2,
   ErrorEmptyState,
   NoDataEmptyState,
   NoTagsEmptyState,
@@ -82,6 +83,13 @@ export { EditTearsheetNarrow } from './EditTearsheetNarrow';
 export { EditFullPage } from './EditFullPage';
 export { EditUpdateCards } from './EditUpdateCards';
 export { Checklist } from './Checklist';
+export { Coachmark } from './Coachmark';
+export { CoachmarkBeacon } from './CoachmarkBeacon';
+export { CoachmarkButton } from './CoachmarkButton';
+export { CoachmarkFixed } from './CoachmarkFixed';
+export { CoachmarkOverlayElement } from './CoachmarkOverlayElement';
+export { CoachmarkOverlayElements } from './CoachmarkOverlayElements';
+export { CoachmarkStack } from './CoachmarkStack';
 export {
   Guidebanner,
   GuidebannerElement,
@@ -90,3 +98,9 @@ export {
 } from './Guidebanner';
 export { InlineTip, InlineTipButton, InlineTipLink } from './InlineTip';
 export { NonLinearReading } from './NonLinearReading';
+export { TruncatedList } from './TruncatedList';
+export { InterstitialScreen } from './InterstitialScreen';
+export { InterstitialScreenView } from './InterstitialScreenView';
+export { InterstitialScreenViewModule } from './InterstitialScreenViewModule';
+export { DelimitedList } from './DelimitedList';
+export { SearchBar } from './SearchBar';