+++ packages/ibm-products-styles/CHANGELOG.md
@@ -3,6 +3,93 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.23.0](https://github.com/carbon-design-system/ibm-products/compare/@carbon/ibm-products-styles@2.22.0...@carbon/ibm-products-styles@2.23.0) (2024-02-06)
+
+
+### Bug Fixes
+
+* **PageHeader:** collapsed breadcrumb tab disappears on resize ([#4068](https://github.com/carbon-design-system/ibm-products/issues/4068)) ([534b5cf](https://github.com/carbon-design-system/ibm-products/commit/534b5cff287ab4da54366f2bc38a2afad5ebf0b0))
+
+
+### Features
+
+* Gradient ai updates ([#4184](https://github.com/carbon-design-system/ibm-products/issues/4184)) ([54aa157](https://github.com/carbon-design-system/ibm-products/commit/54aa15705d613a03361476a522e242d9ff3ea5b6))
+* port Coachmark component to v2 ([#4031](https://github.com/carbon-design-system/ibm-products/issues/4031)) ([dc46cf0](https://github.com/carbon-design-system/ibm-products/commit/dc46cf02ef8c67e98c21fb0887ec704de278bd9a))
+* port InterstitialScreen component to V2 ([#4094](https://github.com/carbon-design-system/ibm-products/issues/4094)) ([827d97d](https://github.com/carbon-design-system/ibm-products/commit/827d97d36adaa4bd8a0b86b6375a78920b1da5a2))
+* **SearchBar:** Migrate SearchBar from Products v1 / Security to Products v2 ([#4202](https://github.com/carbon-design-system/ibm-products/issues/4202)) ([b8653f0](https://github.com/carbon-design-system/ibm-products/commit/b8653f0775708898160e25d55400f22900710f4a))
+
+
+
+
+
+# [2.22.0](https://github.com/carbon-design-system/ibm-products/compare/@carbon/ibm-products-styles@2.21.0...@carbon/ibm-products-styles@2.22.0) (2024-01-31)
+
+
+### Bug Fixes
+
+* side panel and tearsheet fixes ([#4183](https://github.com/carbon-design-system/ibm-products/issues/4183)) ([ee0f3ed](https://github.com/carbon-design-system/ibm-products/commit/ee0f3ed5a2f1ce10aa9850d13854bcf37d07d79a))
+
+
+### Features
+
+* **Datagrid:** add ai slug support to col headers ([#4163](https://github.com/carbon-design-system/ibm-products/issues/4163)) ([f0b17e2](https://github.com/carbon-design-system/ibm-products/commit/f0b17e284ce38157bdd08543c962fb43f4090ad7))
+* **Datagrid:** add expand icon button in filter summary ([#4030](https://github.com/carbon-design-system/ibm-products/issues/4030)) ([bc7641f](https://github.com/carbon-design-system/ibm-products/commit/bc7641f6cab65a1a1d7f759a36185e14f64ad224))
+
+
+
+
+
+# [2.21.0](https://github.com/carbon-design-system/ibm-products/compare/@carbon/ibm-products-styles@2.20.1...@carbon/ibm-products-styles@2.21.0) (2024-01-30)
+
+
+### Bug Fixes
+
+* add min width to 2xl size ([#4019](https://github.com/carbon-design-system/ibm-products/issues/4019)) ([b99ad7c](https://github.com/carbon-design-system/ibm-products/commit/b99ad7cdeea5fad07f474766306f5cd8d1082e5b))
+* **Datagrid:** fix placement of pagination when used with filter panel ([#4152](https://github.com/carbon-design-system/ibm-products/issues/4152)) ([d486005](https://github.com/carbon-design-system/ibm-products/commit/d4860054d31a0a3bbc9e0e1440402fdb8690e8bb))
+* **Datagrid:** sticky row action column solid background ([#4051](https://github.com/carbon-design-system/ibm-products/issues/4051)) ([307c316](https://github.com/carbon-design-system/ibm-products/commit/307c3164629f48741ddeb0739c1333f886e5804e))
+
+
+### Features
+
+* **DelimitedList:** new component DelimitedList ([#4069](https://github.com/carbon-design-system/ibm-products/issues/4069)) ([2e60709](https://github.com/carbon-design-system/ibm-products/commit/2e60709976b08718ec437890e9721e9a1b83500c)), closes [#4047](https://github.com/carbon-design-system/ibm-products/issues/4047)
+* update ai-gradient to phase 2 spec ([#4145](https://github.com/carbon-design-system/ibm-products/issues/4145)) ([4e51d13](https://github.com/carbon-design-system/ibm-products/commit/4e51d132ae60b7aa154eaa433b1a58cdd98fbca2))
+
+
+
+
+
+## [2.20.1](https://github.com/carbon-design-system/ibm-products/compare/@carbon/ibm-products-styles@2.20.0...@carbon/ibm-products-styles@2.20.1) (2024-01-23)
+
+
+### Bug Fixes
+
+* increase title spacing in side panel without subtitle ([#4050](https://github.com/carbon-design-system/ibm-products/issues/4050)) ([cdc3d66](https://github.com/carbon-design-system/ibm-products/commit/cdc3d661d69d2f5889bcc7ea480995bb29e07eae))
+* Potential path issue with CI Check scripts ([#4044](https://github.com/carbon-design-system/ibm-products/issues/4044)) ([73802f9](https://github.com/carbon-design-system/ibm-products/commit/73802f9bd74cfcbb29c39e227f2375c69f1a23e7))
+* **saving:** saving loader alignment ([#4047](https://github.com/carbon-design-system/ibm-products/issues/4047)) ([b6d1d1d](https://github.com/carbon-design-system/ibm-products/commit/b6d1d1da355f8080452ee1f921bf18729cafc326))
+* **TagSet:** margin fix for overflow items ([#4054](https://github.com/carbon-design-system/ibm-products/issues/4054)) ([27406d1](https://github.com/carbon-design-system/ibm-products/commit/27406d116a867b774f39549889f131741d14cd11))
+
+
+
+
+
+# [2.20.0](https://github.com/carbon-design-system/ibm-products/compare/@carbon/ibm-products-styles@2.19.1...@carbon/ibm-products-styles@2.20.0) (2024-01-16)
+
+
+### Bug Fixes
+
+* datagrid empty state hover background color ([#4008](https://github.com/carbon-design-system/ibm-products/issues/4008)) ([21466c8](https://github.com/carbon-design-system/ibm-products/commit/21466c879fd4328e12d5ca1f1abc9d319b27dfa9))
+* **SidePanel:** action buttons layout at 2xl (v2) ([#4000](https://github.com/carbon-design-system/ibm-products/issues/4000)) ([8141984](https://github.com/carbon-design-system/ibm-products/commit/8141984a1d831c36f5b9be78845383293303a936))
+* update format and lint ([#4011](https://github.com/carbon-design-system/ibm-products/issues/4011)) ([8adfa5d](https://github.com/carbon-design-system/ibm-products/commit/8adfa5d49b7d5fac53fe1e336c28ee7ad4f29503))
+
+
+### Features
+
+* **Datagrid:** add responsive updates to filter flyout ([#4001](https://github.com/carbon-design-system/ibm-products/issues/4001)) ([6fa5e6b](https://github.com/carbon-design-system/ibm-products/commit/6fa5e6b12d30718f747f3075155c37b8ff560239))
+
+
+
+
+
 ## [2.19.1](https://github.com/carbon-design-system/ibm-products/compare/@carbon/ibm-products-styles@2.19.0...@carbon/ibm-products-styles@2.19.1) (2023-12-26)
 
 **Note:** Version bump only for package @carbon/ibm-products-styles